Initial start
Added the classic `Position` and `Span` structs. Now we are working on
the `Diagnostic`, which is heavily inspired by the Rustc counterpart.
Neat thing compared to previous attempts, though, is that we should try
to keep the separation we have it now preserved; i.e., `Span`s concern
themselves with fetching source text; `Diagnostic`s do the formatting
and then there are types responsible for creating neat `Diagnostic`s by
providing `Into<Diagnostic>`.