Got something good going
Will make custom Span :') but will also split it into text-formatting mode and binary-formatting mode.

The first is like standard (a snippet), second is a hex grid in case the input is too arbitrary to be parsed as ASCII.

Also, get rekt traits ~ finally seemed to have found a way to do it nicely.