Proposal for new rewrite (WAH)
The problem is: the current implementation of Snack is _almost_ ergonomic. However, it fails at properly assuming error types.

What we need (for multiple reasons) is to streamline the usage of Span's generics to make it wieldable for implementations. Once we do that, we can add bounds to `Combinator`'s generic/associated types such there is a base functionality to expect from all possible parsable types. Then, finally, we can have the combinator macro define generic things for the base case and bim boem bats we have an ergnomic base case. ...is the idea.