Experiment to remove `'s` lifetime from `Spannable`
This is a very controversial decision (among the full 1 developed team).
We have bend over backwards SO OFTEN in the past to keep it, but I think
that, in the current snack ecosystem, it really does more harm than
good. Let's see how it works if we propagate this to the actual parser
impls.